removes unused methods and adds security. to be tested

+++ app/Http/Controllers/UserController.php
@@ -5,42 +5,10 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-
-
-
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
 
     /**
      * Display the specified resource.
@@ -50,18 +18,10 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        if ($user->id !== Auth::id()) {
+            return redirect()->back();
+        };
         return view('user.show', ['user' => $user]);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
     }
 
     /**
@@ -73,6 +33,9 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        if ($user->id !== Auth::id()) {
+            return redirect()->back();
+        };
         $user->name = $request->name;
         $user->email = $request->email;
         $user->save();
@@ -87,6 +50,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if ($user->id !== Auth::id()) {
+            return redirect()->back();
+        };
         $user->delete();
         return redirect('/');
     }
