+++ app/Http/Controllers/WatchlistItemController.php
@@ -7,12 +7,11 @@ use Illuminate\Http\Request;
 
 class WatchlistItemController extends Controller
 {
-    public function index()
-    {
-        return view('watchlistItems.index', ['watchlistItems' => WatchlistItem::all()]);
-    }
     public function destroy(WatchlistItem $watchlistItem)
     {
+        if ($watchlistItem->watchlist()->user()->id !== Auth::id()) {
+            return redirect()->back();
+        };
         $watchlistItem->delete();
         return redirect()->back();
     }
